Refactor Websockets and moved Circuit Breaker
